Made init_hm and hm_entry type agnostic
Working on making the hashmap type agnostic. Right now a user can define the
type that the hashmap stores (int, char, etc.) and the array size of the value
stored. This allows for char arrays of any size alongside storing single ints or
int arrays (these are just examples. In theory any type should be capable of
being stored in any size array).