Hashmap is now type agnostic
Through a void * value variable and refactoring of get_entry, add_entry,
and hm_init, hashmap is type agnostic. It is possible to set the type
and size of the item begin stored through a hm_info struct that is passed to
init_hm (both the key and the value can be any type). Also added a TODO file.